根据新框架，新配置文件，实现各个功能
