+++ config.ini
@@ -1,7 +1,7 @@
 
 [yaml]
-; 所有yaml模板dump的绝对路径的前缀。 例如 /apps/logs/tea-eggs/yaml
-BASICE_PATH=/Users/mi/git/tea-eggs/YAML
+; 所有yaml模板dump的绝对路径的前缀。 例如 /apps/logs/tea-eggs/yaml/
+BASICE_PATH=/Users/mi/git/tea-eggs/YAML/
 [hosts]
 ;使用的host文件，所在位置，绝对路径。
 BASICE_PATH=/Users/mi/git/bck_tea-eggs/taskService/osUser/hosts
@@ -11,7 +11,7 @@ BASICE_PATH=/Users/mi/git/bck_tea-eggs/taskService/osUser/hosts
 SQLALCHEMY_DATABASE_URL=sqlite:////Users/mi/git/tea-eggs/sql_app.db
 
 [ansible]
-become_pass=tea-eggs
-ansible_ssh_user=ops
-ansible_ssh_port=22222
-ansible_ssh_private_key_file=/Users/mi/git/tea-eggs/taskService/test/sshkey/eggs_rsa
+become_pass=123456
+ansible_ssh_user=root
+ansible_ssh_port=12021
+ansible_ssh_private_key_file=/Users/mi/.ssh/id_rsa
