+++ storage/database.py
@@ -12,9 +12,9 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 _project_root = str(pathlib.Path(__file__).resolve().parents[1])
 sys.path.append(_project_root)
-from config import DATABASE
+from config import DATABASE_URL
 
-SQLALCHEMY_DATABASE_URL = DATABASE
+SQLALCHEMY_DATABASE_URL = DATABASE_URL
 # SQLALCHEMY_DATABASE_URL = "postgresql://user:password@postgresserver/db"
 
 engine = create_engine(
