+++ taskService/test/test_task.py
@@ -12,9 +12,9 @@ sys.path.append(_project_root)
 from traceback import print_exc
 
 
-from taskService.osUser.new_lock_user import ansibleLockUserTask
+from taskService.osUser.os_user_tasks import ansibleLockUserTask
 from storage.osUser.models import LockUserTask
-
+from config import ANSIBLE_VARS
 
 def test_lock_user_task_01():
     """
@@ -23,14 +23,7 @@ def test_lock_user_task_01():
     docker-compose up 
     ping  172.20.16.2 能通过时再执行测试
     """
-    
-    ansibile_vars={ 
-    'ansible_ssh_user' : 'ops',
-    'ansible_ssh_port' : '22222',
-    'ansible_ssh_private_key_file' : "/git/tea-eggs/taskService/test/sshkey/eggs_rsa"
-    }
-    runtime_task=ansibleLockUserTask(become_pass="tea-eggs",yaml_save_path=_project_root+"/tmp.yml",ansibile_vars=ansibile_vars)
-
+    runtime_task=ansibleLockUserTask(become_pass="tea-eggs",yaml_save_path=_project_root+"/TEST/lock_user.yml",ansible_vars=ANSIBLE_VARS)
     task_model=LockUserTask()
     task_model.host="127.0.0.1"
     task_model.username="ops"
